Refactor safety things to same module, add annotator arbiter.
