+++ src/modelplane/evaluator/outputs.py
@@ -9,17 +9,3 @@ class Output:
 
     def __repr__(self) -> str:
         return f"{self.name} ({self.__class__.__name__})"
-
-
-class Safety(Output):
-
-    def __init__(self, is_safe: bool) -> None:
-        self.is_safe = is_safe
-
-    @property
-    def name(self) -> str:
-        return "SAFE" if self.is_safe else "UNSAFE"
-
-
-SAFE = Safety(is_safe=True)
-UNSAFE = Safety(is_safe=False)
